add attestation script
Small script that uses gpg and freetsa to perform an attestation of a
statement.

Usage:
```
  ./attest.sh --attest <gpg-key-id> "message text"
  ./attest.sh --verify <attestation.tgz>
```

The script signed the message text with the GPG key of *someone* and
fires an attestation request of the signature to freetsa.org that then
signs and attests the signature thus asserting the timestamp.

Depending on how much you trust the coupling of the GPG key to an
individual and how much trust freetsa as a service this should provide
pretty hard proof that *someone* asserted the "message text" claim at a
certain time.